+++ units/Cargo.toml
@@ -20,12 +20,13 @@ alloc = ["internals/alloc","serde?/alloc"]
 [dependencies]
 internals = { package = "bitcoin-internals", path = "../internals" }
 
-serde = { version = "1.0.103", default-features = false, features = ["derive"], optional = true }
+serde = { version = "1.0.103", default-features = false, optional = true }
 arbitrary = { version = "1.4", optional = true }
 
 [dev-dependencies]
 internals = { package = "bitcoin-internals", path = "../internals", features = ["test-serde"] }
 bincode = "1.3.1"
+serde = { version = "1.0.103", default-features = false, features = ["derive"] }
 serde_test = "1.0"
 serde_json = "1.0"
